Agentic/Kenton/pyproject.toml
[tool.poetry]
name = "deep-research-agent"
version = "0.1.0"
description = "Deep Research AI agent using OpenAI Agents SDK"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
openai-agents = "^0.0.15"
httpx = "^0.27.0"
requests = "^2.28.0"
python-dotenv = "^1.0.0"
fastapi = "^0.115.12"
uvicorn = {extras = ["standard"], version = "^0.34.2"}
sse-starlette = "^2.3.5"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
ruff = "^0.0.290"
black = "^23.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"